day 9 done

+++ src/main/kotlin/ro/codespace/aoc/day7/Day7a.kt
@@ -16,7 +16,7 @@ fun computeSignal(configuration: List<Int>, code: List<Int>): Int {
         reader.send(signal)
         program()
 
-        signal = program.lastOutput
+        signal = program.lastOutput.toInt()
     }
     return signal
 }
